Add support for OS upgrades and updates in package managers
This commit adds support for checking and upgrading the operating system
in package managers. The PackageManager interface has been extended with
two new methods, CheckOSUpdates and UpgradeOS. Additionally, a new
struct Update has been defined to hold package update information.

The PackageManager interface now includes the following methods:

CheckOSUpdates: This method checks for available updates for the
operating system on the given host and returns a list of update strings.

UpgradeOS: This method performs the OS upgrade on the given host and
returns a list of Update objects representing the packages that were
upgraded.

The implementation of these methods has been provided for the
YumPackageManager, AptPackageManager, and BrewPackageManager.

The parseUpdates function has been added to help parse the output of
package manager commands and extract the package name and version for
each update.

These changes enable the package managers to handle OS upgrades and
updates in addition to package-specific operations.