Add SSHClient as a dependency to UnixHost
This commit introduces the SSHClient interface in the steelcut package,
which provides an abstraction over the SSH dial functionality. The
actual implementation (SSHClientImpl) is defined in the main.go file,
which uses the x/crypto/ssh.Dial function.

The UnixHost structure has been extended with a new field, SSHClient, to
hold an instance of a type implementing the SSHClient interface. This
enables dependency injection of the SSH client, which is used in the
RunCommand method.

The main function has been updated to create an instance of
SSHClientImpl and inject it into the UnixHost via the newly added
WithSSHClient function.

This change allows for more flexibility and testability in terms of how
SSH connections are established.