Add CopyFile method to UnixHost in steelcut package
This commit introduces a new method `CopyFile` to the `UnixHost` type
within the steelcut package. This method allows for copying a file from
a local path to a remote path over SSH using the SCP protocol. Necessary
imports were added to support this functionality, including "io", "os",
and "path/filepath".

This enhances the capability of the UnixHost type to handle file
transfers in addition to command execution, providing more comprehensive
support for remote operations.